Also fix an issue in retry logic
After changing the order of deserialization policy and retry policy,
`error.code` is now populated properly by deserialization policy. This
surfaces an issue where an error with code `ResourceNotFound` will
also be retried because it contains `eNotFound` and we use
`error.code.toString().toUpperCase().includes()` to see if the error
is in the list. It passed the check for the network error code
`ENOUTFOUND`.  This change fixes it by using exact match when checking
error code.