[core-http] use response status to determine whether body is stream
Our current code has the incorrect assumption that if any body mapper
of an operation spec is of `Stream` type then response body must be a
stream and `httpRequest.streamResponseBody` is set according to this
assumption. However, Storage issue #12997 shows that the default
response of Download Blob operation does not have a stream body with
the following operation spec:

```typescript
  responses: {
    200: {
      bodyMapper: {
        serializedName: "parsedResponse",
        type: {
          name: "Stream"
        }
      },
      headersMapper: Mappers.BlobDownloadHeaders
    },
    ...
    default: {
      bodyMapper: Mappers.StorageError,
      headersMapper: Mappers.BlobDownloadHeaders
    }
  },
```
Treating response body as stream in the default case prevent us from
parsing the body which leads to missing `code` and `message` from the
error object.

This change fixes the issue by actually checking the body type of the
response corresponding to the response status code.