Deprecate WebResource{Like}.streamResponseBody
because a boolean value is not enough to determine whether a response
is streaming or not. A list of streaming response status code
`streamResponseStatusCodes` is introduced and used instead.

ServiceClient no longer sets the deprecated
`streamResponsebody` property based on incorrect assumption. However
we still respect `streamResponseBody` in case it's set by users.

Core v2 removes `streamResponseBody`.