[core-http] reduce memory usage of proxy policy cache
Currently we cache the result of whether a request to a url should bypass proxy
using url as the cache key. In some scenarios the number of unique request urls
can be very large, for example, when iterating through thousands of Azure
Storage blob items and retrieving their properties.

This PR changes the cache key to host names. Code is simplified a bit and
updated to allow more reliable testing.  The fix is also ported to Core v2.