[core-http] delay saving of DOM parser error namespace
We eagerly save the namespace of "parsererror" element on module load. It's
unnecessary because in most scenarios we would not encounter a parse error.
Furthermore, it caused Content Security Policy error in Chrome version 86.0 or
newer because the returned DOM object for parser errors contains some CSS
styles.

This change fixes the issue by lazily loading the error namespace when the
parsed DOM document contains "parsererror" elements.

Fixes #13268