Updated the Sample Code block
Suggesting an alternative Sample Code block that shows using Service Principal login as well as listing Asset names in the account instead. 

This sample requires a .env file with values from the portal, so I noted that as well.  I'll try to get all of this into a sample project to link to soon also. 
@ramya-rao-a