[Storage-Blob] Fix error listing blobs with "_" metadata key name (#12421)
The cause of the issue is that xml parser in core-http was using
hard-coded "_" as the key to access parsed xml element content. Now
that core-http added support to customzie the XML_CHARKEY, fix on
Storage side simply pass "#" as the xml char key to use. Note that "#"
is not allowed in storage metadata keys so it's safe to use as the key
to access in xml parser.

* re-generate recording to remove `sr` query parameter

* ignore sip

* Add CHANGELOG entry

Co-authored-by: Lin Jian <1215122919@qq.com>