[Storage] retry on incomplete XML responses (#13076)
* [Storage] retry on incomplete XML responses

When service times out (default max 30s) it terminates the connection
but current stable version of `node_fetch` doesn't report
error. Instead it returns the incomplete response which leads to XML
parse error.  It's unlikely that service would send back incomplete
response on purpose so it doesn't hurt to treat this error as a
`TIMEOUT` error and retry the request.

The deserialization policy factory needs to move below retry policy
factory so parse error from deserialization can be retried.

* Add browser recording for new test

* Verify that error has been injected

* Tweak comments

* Apply same fix to file-share/file-datalake/queue

* Also fix an issue in retry logic

After changing the order of deserialization policy and retry policy,
`error.code` is now populated properly by deserialization policy. This
surfaces an issue where an error with code `ResourceNotFound` will
also be retried because it contains `eNotFound` and we use
`error.code.toString().toUpperCase().includes()` to see if the error
is in the list. It passed the check for the network error code
`ENOUTFOUND`.  This change fixes it by using exact match when checking
error code.

* Add CHANGELOG entries