Remove the `hidden` attribute for all spots, use `internal` only. Also remove the tsdoc.json as it was only there to allow for the usage of the `hidden` attribute. (#13469)
Remove the `hidden` attribute for all spots, use `internal` only. Also remove the tsdoc.json as it was only there to allow for the usage of the `hidden` attribute.

NOTE: I've also changed the generated clients. Next time we use the official autorest generator it should generate the exact same thing so this shouldn't be an issue.

Fixes #13445