[Recorder] Workaround for the "url-path with single quotes" bug in nock (#13474)
## Problem
As part of moving `data-tables` from `core-http` to `core-https` at #12548, @xirzec has observed that the generated recordings for the node tests was invalid if the url-path has single quotes.
Upon investigating, I could see that the request url at core-http and core-https levels is the same.
The problem is seen because of the switch from `node-fetch`(core-http) to native `https`(core-https).

`node-fetch` encodes the request before sending which encodes the single quotes to `%27` and nock was able to save the fixtures with the encoded request nicely.
When migrated to `https` in core-https, there is a difference in behavior with the URL path, there is no default encoding, leading to the presence of single quotes.

## Bug(in `nock`)
Given the above problem, "nock" does capture the request with single quotes in the URL path, but fails to save the fixture properly leading to invalid recordings on our end.
Reference -  https://github.com/nock/nock/blob/dc04694dcd3186a396e48567bdfabbdad2761f6a/lib/recorder.js#L134

This line 
> lines.push(\`  .${methodName}('${path}', ${JSON.stringify(requestBody)})\`)

Single quotes can be present in the url path though unusual. When the url path has single quotes, the fixture generated by "nock" is incorrect since it doesn't consider the case. 
#### Examples below
- `.delete('/Tables('node')')` 
- `.get('/Tables('node')')` 
- `.post('/Tables('node')', {"TableName":"testTablenode"})`

## Fix
To avoid this problem, we replace the single quotes surrounding the url-path in the recording with backticks(`) if there are any single quotes present in the url-path. This would fix the invalid recordings.
This is backward compatible as well.
This would work for both core-http and core-https.

## To Do
- [x] Add workaround
- [x] Tests for the workaround
- [x] Create an issue in the "nock" repo 
    - [x] https://github.com/nock/nock/issues/2136
- [x] Changelog