[Recorder] Single quotes in the url-path - careful revert of the reverted "workaround" (#13494)
https://github.com/Azure/azure-sdk-for-js/pull/13474 added a workaround in the recorder but was reverted soon at https://github.com/Azure/azure-sdk-for-js/pull/13488 because the presence of single quotes in the request body had corrupted the updated fixture due to the workaround.

For the case with request bodies, since the request body may contain literally anything, it is not trivial to come up with a regex to split the arguments. Any regex would have assumptions on the line saved by nock, i.e., the number of spaces in between the arguments and the characters with which the payload would start and what the url-path may contain.

Given the problems, it might be much better if we get the fix from nock before the string/fixture is saved.

And since we have only seen the single quotes with the delete requests and that too for the requests without the request body, I'm inclined to have the workaround only for that case(without the request bodies).

I have left a TODO for the case with request bodies for now. 

Have tested this fix with the data-tables SDK, and I was able to record and playback all the `data-tables` tests. This would unblock #12548.