[core-http] reduce memory usage of proxy policy cache (#13329)
Currently we cache the result of whether a request to a url should bypass proxy
using url as the cache key. In some scenarios the number of unique request urls
can be very large, for example, when iterating through thousands of Azure
Storage blob items and retrieving their properties.

This PR changes the cache key to host names. Code is simplified a bit and
updated to allow more reliable testing.  The fix is also ported to Core v2.

* Align with .NET Azure.Core implementation

Our code was ported from
https://github.com/Azure/azure-sdk-for-net/blob/8cca811371159e527159c7eb65602477898683e2/sdk/core/Azure.Core/src/Pipeline/Internal/HttpEnvironmentProxy.cs#L210
but some differences were introduced. Change to match.