[Text Analytics] Integrating action failures and successes into one array (#13568)
The text analytics crew recently agreed on treating action errors in a slightly different way and this PR implements this change. The `beginAnalyzeBatchActions` takes a list of actions (previously known as tasks) as input and apply all of them at once to the input documents and return the results. The results could be a mix of action errors and succeeded action results but the response keeps them separate. This PR combines both types of results into one list of action results (similar to how results on the document level are presented).

Things to note:
- If the service returns failed as a the job status, we used to create an exception out of all errors in the top-level errors list and throw it. Instead, in this PR, I no longer throw an exception and just return those errors as items in the action results list. This approach has the advantage of not confusing the user by using two different ways to show the same type of errors (exceptions and the result list). It also makes it easier for the user to inspect the errors by having them as individual items in a list rather than being one big string in an exception. 
- The list of action results is sorted with respect to the input list of actions.
- I commented all new functions and types so hopefully that will make it easier to review.
- I also bumped the package version.
- ~I will add new tests in another PR.~ I added unit test cases and one client test case.
- Code complete is February 5th so I would really appreciate your review before end of the day.