[Identity] DefaultAzureCredential clarification
This PR is intended to help portray a better message to our customers about the DefaultAzureCredential. This credential is intended to be used to get things started, and not as the universal way to go over time. I took this text from something Charles Lowell sent me on Teams 😁 I loved the phrasing!

Related to: #13799