Remove paths from container/blob operation specs
Storage has a baseUri of "{url}" and path of "{containerName}" or
"{containerName}/{blob}". Though in reality the two paths are not
used (there are also no replacement values passed for them from
options). Instead we relied on a behavior of `URLBuilder` that when
"{url}" is replaced with a url that contains a path,
"{url}/{containerName}" or "{url}/{containerName}/{blob}" would lose
their path after the first "/".  Storage always pass a full url to
"{url}" so appending "{containerName}" or "{containerName}/{blob}"
would cause problems.

The swagger probably needs some fix.

This change removes them from the generated code for now.