[KeyVault] - Replace `keyId: URL` with `certificateKeyId: string`  (#13908)
## what

- Deprecate `secretProperties.keyId` and replace it with `secretProperties.certificateKeyId`
- Remove `dom` from tsconfig

## why

`secretProperties.keyId` being a URL is the only reason we need `dom` lib in this library.
Even though it has never been populated we cannot remove it because we might break 
someone's compilation (although it would have never worked at runtime). 

Instead we'll deprecate it, note that it will always be undefined, and make it an any so
we can remove the dependency on `dom`. 

Refer to #9639 for previous work in this area.

Resolves #10726