[abort-controller] listeners that self-remove from abortSignal when invoked no longer prevent other listeners from being invoked (#13987)
Fixes #13985

See the issue linked above for a RCA on the bug this fixes.

## Issue
When using the @azure/abort-controller package, attaching event listeners to an abortSignal that remove themselves after being invoked can prevent other listeners attached to the same signal from being invoked.

## Solution
With this update, when abort() is called, we create a copy of the listeners array that isn't mutated when removeListener() is called on the abortSignal. This ensures that listeners are invoked in the same order they were added.