[EventGrid] Add Performance Test (#14091)
The EventGrid library is a fairly lightweight wrapper over the
service, but we would like to have a least one performance test so we
have a baseline to compare things with going forward.

This test just sends events using the CloudEvents schema to an
endpoint. This operation was chosen because it has the most work
being done by the SDK (there's a small level of translation that needs
to be done to adapt the user facing model to what the service expects
on the wire).