[EventGrid] Add test to lock in endpoint behavior
We require that the user inclide the `/api/events` path as part of the
endpoint when creating a client. This is so that we remain flexable
and allow the event grid service to evolve the structure of their
endpoint URLs.

We already had this behavior, but so just add a test to lock the new
behavior in.