Upgrading to opentelemetry 1.0.0 (rc.0)
Did a few things that made this MUCH easier.

Now that everyone is using the createSpan from @azure/core-tracing we
no longer need _every_ project to reference opentelemetry/api! That has
been removed as part of this PR.

Unfortunately, the leaky nature of JS means that packages still need to
worry about opentelemetry when they build their browser bundle for
testing. To make that simpler I've added a common function to dev-tool
that everyone can call in their rollup that will give them the correct
named exports. This is hooked up for everyone at this point, so the next
time something like this happens I should be able to control it
centrally.

Now for the API breaking changes that I had to fix:
- CanonicalCode is gone and is replaced with SpanStatusCode.
  SpanStatusCode has a much smaller set of codes (literally: ERROR, OK
  or UNSET) so that simplified most of the way we were handling setting
  a span status on error.
- There is a new field (`tracingContext`) that contains `Context`. You
  now pass a context, not a span, to indicate what your "parent" is.
  You'll see this where I've removed `SpanOptions.parentSpan`. Mostly
  it's a simple replacement.