Cosmos: project depends on azure/identity
The package.json for the project needs to have the identity package
within its dependencies, not the dev dependencies, or else it will
cause problems when building in TypeScript projects.

This:

- moves the package from dev deps to deps
- moves a types package from deps to dev deps

Note, there is reference to a library called @azure/dev-tool that
is not published, so I am unable to run tests.

See also:
- https://github.com/Azure/azure-sdk-for-js/issues/13202
- https://github.com/Azure/azure-sdk-for-js/issues/14236