[Text Analytics] Fix samples (#14255)
This PR does the following:
- updates samples so the typescript ones can run with `node8` with no issues by setting the compilation target to `ES6`.
- fixes the recognize PII sample where a the behavior for the protected healthcare information parameter changed in the service. This PR updates the input text used with this parameter to one that does not have protected healthcare information according to the new update. More information can be found [here](https://docs.microsoft.com/en-us/azure/cognitive-services/text-analytics/named-entity-types?tabs=personal#government-and-countryregion-specific-identification). Also, this is the output of the updated sample:
```
Running recognizePii sample

The redacted text is "My phone number is ********" and found the following PII entities
       - "555-5555" of type PhoneNumber
There are no PHI entities in this text: "His EU passport number is X65097105"
Also there is nothing to redact: "His EU passport number is X65097105"
But there are other PII entities in that text:
       - "X65097105" of type EUPassportNumber
You can choose to get SSN entities only, or any other PII category or a combination of them. For example, in this text: "Patient name is Joe and SSN is 859-98-0987", this is the SSN number:
       - "859-98-0987"
```
- adds a new npm script `execute:sample` to run individual samples.
- fixes a bug in the javascript sample for beginAnalyzeHealthcareEntities.
- re-enable text analytics samples in smoke tests.

The javascript samples still fail with `node8` because of the lack of support for `ES6`, but this is fine because we should always write our samples using the recent features of the language and we do not want them to look outdated.