[EventGrid] Add Additional ACS Events, Address Arch-Board Feedback for GA (#14136)
* [EventGrid] Regenerate Code

Re-run the code generation based on the latest README.md in the rest
specs repository. This picks up some changes to some Azure
Communication Services event types that the service has changed.

* [EventGrid] Updates for GA

Adopt feedback from architecture board review, the changes here are
focused on the consumption side:

- We decided that we did not want to apply AutoRest style mapping for
 system event data payloads, instead we would keep the `data` value as
 is. This allows us to remove a bunch of our custom logic around
 invoking the mappers. One implication here is that we need to remove
 the `format` directive for model types in our swagger before
 generating code with auto-rest, so that the interfaces defined for
 custom events use `string` as the type instead of `Date`.

- Some light renaming of interface names tied to System Event
  Data. There was a common convention from some services to use a
  `Properties` suffix in their schema name, which caused us to
  generate interfaces with the suffix `Properties` which was not very
  idiomatic.

- Mark properties in shapes from System Events as present, based on
  the sample data that is published by the EventGrid service team on
  docs.microsoft.com.

Fixes #10992

* [EventGrid] Add test to lock in endpoint behavior

We require that the user inclide the `/api/events` path as part of the
endpoint when creating a client. This is so that we remain flexable
and allow the event grid service to evolve the structure of their
endpoint URLs.

We already had this behavior, but so just add a test to lock the new
behavior in.

* [EventGrid] Update Recordings