adding a check-format step in CI (#13983)
Adds a new analyze step to run `rushx check-format` in the package the opened PR is against. We are doing this after killing the git precommit hook that used to run the formatter here: https://github.com/Azure/azure-sdk-for-js/pull/13982. I am not very familiar with engineering systems scripts so I might have missed something.