[EventGrid] Mappings for new ACS Events (#14410)
* [EventGrid] Fix ACS Event Names

The Azure Communication Services team noticed that some of their event
shapes were wrong and have [updated the
swagger](https://github.com/Azure/azure-rest-api-specs/pull/13485) to
address this.

This commit pulls these changes into our SDK.

Fixes #14345

* [EventGrid] Add types for  RecordingFileStatusUpdated event

This is a new event ACS is sending.

* [EventGrid] Remove incorrect ACS System Events

After discussion, we are comfortable removing these two event names
from our mapping. The rationale here is that the service never sent
events using these names (they made a typo when documenting the event
names in Swagger) and so any code using them was going to be wrong. In
this case, we like that if you're using TypeScript you'll see a
compile time issue here because it will be pointing to place in your
code where things were never going to behave as you might have
expected.

The `CHANGELOG.md` has been updated to provide a little more
perscriptive guidence on what to do here, and we feel OK about not
doing a major version bump.