[Identity] DeviceCodeCredential constructor unification (#14443)
DeviceCodeCredential stopped having required parameters a while ago, when we started completing the values for tenant ID and client ID by default. However, even though our guidelines establish that we should group all of the optional parameters in option bags, the signature of this constructor never changed.

As part of my recent Identity changes, I've made it so the DeviceCodeCredential constructor could also accept only receiving the optional parameters. The intention there was to not do a breaking change initially, but since we're targeting a major version release, having this constructor overload feels unnecessary.

This PR makes it so DeviceCodeCredential only accepts an options bag as the first parameter, without overloads.

While doing these changes, I was able to spot a bug I introduced in my recent changes (currently on master, but not released) by which clientId and tenantId were not properly passed through if sent only through the options bag. While this is fixed by this PR too, that fix caused the recordings not to match, as some of the recordings were using the default values for clientId and tenantId instead of the ones passed through. I've fixed the tests to match what the recordings represented, which essentially allows us to test that the default values for tenantId and clientId can work, meaning that now I'm treating the device code tests as if sometimes used the default parameters.

Fixes: #14420