[Identity] Making MSAL extensions an unlisted dependency (#14418)
`@azure/msal-node-extensions` has some very inconvenient requirements for our developers and for our community. It requires (besides Python in the three environments):

- On Linux, to have installed `libsecret` and `build-essentials`.
- On OSX, to have XCode configured.
- On Windows: To have Visual Studio Code installed with C++ (this is reportedly about 60gb)

While we move this dependency (and KeyTar) to a new package, this PR makes it so the dependency is "unlisted": If users try to use this feature, we will throw an error asking them to install the dependency. Tests that use this feature get skipped if the dependency is not present, and run if the dependency is present.

I will add more clarity in the docs as part of this other PR: https://github.com/Azure/azure-sdk-for-js/pull/14405