[EventGrid] Move to Samples v2 Framework
This change moves our samples to be generated via `devtool`, as part
of the samples quality effort. Much of the diff here is just moving
code around to confirm to the new patterns, as outlined in the
migration guide, but one interesting outcome is that since the samples
are now part of the ts compliation for the package itself, and our
samples use service bus (which itself uses AsyncIterators) we have to
add a `lib` in our `tsconfig.json`.

We do not add any polyfill or anything like that to our `package.json`
because our implementation does not use these features.

After talking about this with will, I did manually patch the
*generated* tsconfig.json for our samples to say we target ES2018
(since the samples will need this), and he will add support for that
for the tool going forward.

In addition, we add some more resources to our `test-resources.json`
so that these samples can run in CI and be validated.

Fixes #14471
Fixes #14573