Fix appendPath() issue when pathToAppend contains `?`
There's an issue in appending to path when the value to append
contains `?`. For example,

```ts
  const parsedUrl = new URL("https://example.com");
  parsedUrl.pathname = "/path?abc%3Ddef";
```

will cause the question mark to be encoded, then `parsedUrl.toString()`
would give us unexpected `'https://example.com/path%3Fabc%3Ddef'`

This PR fixes the issue by appending path and search part separately
when `?` is included in the value to append to path.