[Identity] No more inMemoryPersistence!
Now that we've released 2.0.0-beta.1 and the Identity hotfix that I
wanted to release, I was able to sit down and start writing a sample
showcasing some of the "bugs" I found. This helped me bounce against
the code and eventually figure out that it was in fact _me_ who had
written a bug!

This if:

```ts
if (this.publicApp && this.confidentialApp) {
  return;
}
```

in the code that defines the configuration for Node.js was causing the
MSAL client to be defined multiple times, rendering the in-memory cache
as useless.

The solution? To switch `&&` to `||`.

It seems obvious in retrospective!

This means that there's no need neither for the `inMemoryCache`, nor for
the line that I used to "read" from the cache. Therefore:

Fixes #14372
Fixes #14373