[dev-tool] Added declarative spec for extra files (#14576)
This adds a little feature to the samples publication system that both Event Hubs and Form Recognizer need. It allows us to copy extra files into the output directory using an `extraFiles` field in the sample configuration.

`extraFiles` is a `Record<string, string[]>` that specifies a map from source file to an array of destinations to which that file should be copied. For example, if you have an extra sample directory (for browser samples or any other special use case), you can specify it like this:

```javascript
{
  "//sampleConfiguration": {
    ...,
    "extraFiles": {
      "./samples-browser": ["browser"]
    },
    ...
  }
}
```

or to copy static assets:

```javascript
{
  "//sampleConfiguration": {
    ...,
    "extraFiles": {
      "./assets": ["javascript/assets", "typescript/assets"]
    },
    ...
  }
}
```