[dev-tool] Sample README snippet support (#14591)
This PR adds support for adding inline markdown snippets to the generated sample README files. This support is exposed through `//sampleConfiguration` via the `customSnippets` field. It is represented as a map from snippet name to snippet file. There are currently three supported snippet names:

- `header` (top of the file, useful for important notices)
- `prerequisites` (just after the "required resources" section, useful for describing required setup)
- `footer` (bottom of the file, useful for notes)

Unspecified snippets are omitted.

We are adding this feature because Key Vault samples need a way to describe the process of enabling soft-delete on the keyvault resource and because some samples need a way to add notices about service version support to the top of the file.

Example:

```javascript
{
  ...,
  "//sampleConfiguration": {
    ...,
    "customSnippets": {
      "prerequisites": "samples-dev/snippets/_prerequisites.md"
    },
    ...
  },
  ...
}
```

With the example configuration above, the _contents_ of `sdk/<service>/<package>/samples-dev/snippets/_prerequisites.md` will be inserted into the output README files (in both the JavaScript and TypeScript packages).

Incidentally, I also changed the default target to ES2018, since we're targeting Node 12 for samples. This will technically support Node 10, but it allows the compiler to skip downleveling of async iteration. No samples that are already generated need to regenerate for the sake of just this change.