[Identity] Dropping Implicit Auth flow (#14558)
As part of `2.0.0-beta.1` we released with support for both authentication flows, Implicit and Auth Code. We'd rather release `2.0.0` without Implicit Auth, because this authentication flow is deprecated, and to also drop the `msal` dependency in the process. We're also interested in having an easy way to add this feature back, in case it's needed.

This PR does exactly that:
- Drops Implicit Auth Flow support.
- Drops the `msal` dependency.
- Updates the manual tests and docs.

The changes should be easy to re-do if we ever need to add this auth flow back.

**Background:**  
Auth Code was thought to be a backwards compatible change until we verified that some of the AAD apps will not be able to work both with Implicit Auth and Auth Code. From that point, we decided to take the opportunity to do a major version change and do improvements across the board. In the process, we realized that the major version change was giving us the opportunity to drop Implicit Auth completely. Our main concern was that, if in the future any customer needed us to support Implicit Auth after 2.0.0 is released, it should be easy to add it back.

Fixes #14276