[Identity] InteractiveBrowserCredential should handle open() throws (#14617)
This PR adds a way to handle throws coming form the `open` dependency. One way this dependency will throw is if the are no browsers available. One quick way to test this is to install `open` in a Linux VM and to run this Node code: 

```js
const open = require("open");

(async () => open("http://sadasant.com/", { wait: true }))()
  .then(console.log)
  .catch(e => console.log("open() error", { message: e.message }));
```

Which will cause this error:

```
open() error { message: 'spawn xdg-open ENOENT' }
```

This effectively means that InteractiveBrowserCredential  will throw if there are no browsers available. This could be useful if we ever want to provide a backup to DeviceCode on those cases.

Though this issue is not blocking the release, introducing InteractiveBrowserCredential tests to our CI will help us catch anything we haven't noticed yet.

Fixes #14616