[KeyVault] - Add multi-version tests for KeyVault Keys (#14587)
## What

- Adds support for live testing multiple service versions
- Adds the matrix config and plumbing to pass service version as an environment variable
- Adds a helper method to support skipping tests when the service version isn't supported

## Why

KeyVault currently supports three service versions: 7.0, 7.1, and 7.2. Being able to ensure our
live tests run against previous service versions is a way to future proof ourselves and allow us
to ensure continued support.

After a bit of discussion we agreed that it would be worthwhile to add the indirection of passing
the service version as an environment variable and avoid having to run all service versions tests
in sequence.