[identity] 1.3.0 hotfix: Update to core-tracing 1.0.0-preview.11 #14649 (#14682)
This PR updates the current GA version of identity to use the new core-tracing libraries, which will bring all released packages up to the same version of opentelemetry/api. Without this we'd end up in a situation where multiple versions of the API packages are in-use, which could lead to lost/unparented telemetry.

As part of this PR I've also had to remove (from rush.json) all the dependencies that were coming from within the branch since they are outdated. This means that identity will build clean, using only published dependencies, which is what we intend.