[KeyVault ] - Fix Managed HSM access control live test (#14779)
A recent change in MHSM required us to add the Crypto User role to our existing user. Unfortunately that's also the role we use to test role assignments. 

This PR just changes the role we use for testing to something that will likely not be needed and unblocks live tests.