Fixing session peeking test that was failing randomly.
The crux is that subscriptions and queues behave differently when messages are not present for a session. Queues won't even let you create a receiver, so just have a receiver is enough to know that a message exists.

Subscriptions will let you create the receiver even if no messages exist, in which case an immediate peek() will result in no messages (correctly).

This commit makes it so we run the same peek tests against queues and subscriptions and also inserts a receiveMessages() call to ensure that messages exist before we actually do the peek(). (receiveMessages should always wait long enough for a message to arrive).