[Service Bus] Bug Fix: Unexpected `expiresAtUtc: Invalid Date` in the receivedMessage (#13543)
Fixes #12883 
## Bug
"Invalid date" for the `expiresAtUtc` property
```
azure:service-bus:messages:verbose AmqpMessage to ServiceBusReceivedMessage: {
  _amqpAnnotatedMessage: {
    header: { deliveryCount: 0 },
  .
  .
  .
  expiresAtUtc: Invalid Date,
  .
  .
  .
}
```

## Background
"ttl" on the received message will be undefined if the sent message doesn't have the "ttl" property set to a value. 
In this case, the service assumes the default infinite value for expiry time.
Reference: https://docs.microsoft.com/en-us/azure/service-bus-messaging/message-expiration#entity-level-expiration

### Cause (expiresAtUtc)
In the SDK, for a received message, we calculate the expiresAtUtc using the "ttl" and the enqueued time.
- If the msg.ttl > max_duration - msg.enqueuedTime, calculate msg.expiresAtUtc with max_duration
- If the msg.ttl < max_duration - msg.enqueuedTime, calculate msg.expiresAtUtc with msg.enqueuedTime  + msg.ttl
- If the msg.ttl is not defined on the received message, we return Invalid Date for the expiresAtUtc property.

## .NET SDK
TTL is defaulted to TimeSpan.MaxValue in .NET.
For ExpiresAt calculation:
- if AbsoluteExpiryTime is populated we use that.
- if TTL > maxDateTime - EnqueuedTime, use TimeSpan.MaxValue
- Else use EnqueuedTime + TTL


## Fix for JS SDK
If the msg.ttl is not defined, set it to the max value as the service says that. This allows the proper calculation of `expiresAtUtc`.

## TODO
- [x] Bug Fix
- [x] Test
- [x] Changelog